Move hooks to a separate file
