+++ lib/AsyncStorage.js
@@ -42,7 +42,7 @@ type MultiRequest = {|
  *
  * See http://facebook.github.io/react-native/docs/asyncstorage.html
  */
-const AsyncStorage = {
+export default {
   _getRequests: ([]: Array<MultiRequest>),
   _getKeys: ([]: Array<string>),
   _immediate: (null: ?number),
@@ -366,29 +366,3 @@ function convertError(error): ?Error {
   out.key = error.key;
   return out;
 }
-
-type AsyncStorageHook = {
-  getItem: (
-    callback?: ?(error: ?Error, result: string | null) => void,
-  ) => Promise<string | null>,
-  setItem: (
-    value: string,
-    callback?: ?(error: ?Error) => void,
-  ) => Promise<null>,
-  mergeItem: (
-    value: string,
-    callback?: ?(error: ?Error) => void,
-  ) => Promise<null>,
-  removeItem: (callback?: ?(error: ?Error) => void) => Promise<null>,
-};
-
-export function useAsyncStorage(key: string): AsyncStorageHook {
-  return {
-    getItem: (...args) => AsyncStorage.getItem(key, ...args),
-    setItem: (...args) => AsyncStorage.setItem(key, ...args),
-    mergeItem: (...args) => AsyncStorage.mergeItem(key, ...args),
-    removeItem: (...args) => AsyncStorage.removeItem(key, ...args),
-  };
-}
-
-export default AsyncStorage;
