+++ lib/index.js
@@ -1,0 +1,4 @@
+import AsyncStorage from './AsyncStorage';
+
+export { useAsyncStorage } from './hooks'
+export default AsyncStorage;
